% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-by-two.R
\name{mle_lor}
\alias{mle_lor}
\title{Inference for MLE of common log odds ratios in meta-analysis of two-by-two tables}
\usage{
mle_lor(
  ai,
  n1i,
  ci,
  n2i,
  correction_factor = 0.1,
  alternative = "two.sided",
  conf.level = 0.95
)
}
\arguments{
\item{ai:}{number of responders in group 1}

\item{n1i:}{sample size of group 1}

\item{ci:}{number of responders in group 2}

\item{n2i:}{sample size of group 2}
}
\description{
Inference for MLE of common log odds ratios in meta-analysis of two-by-two tables
}
