% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{overall_proportion}
\alias{overall_proportion}
\title{Inference for overall proportion in proportion meta-analysis}
\usage{
overall_proportion(xi, ni, alternative = "two.sided", conf.level = 0.95, ...)
}
\arguments{
\item{xi:}{number of responders in each study}

\item{ni:}{study sample size}
}
\description{
Inference for overall proportion in proportion meta-analysis
}
