% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-by-two.R
\name{pwa_lrr}
\alias{pwa_lrr}
\title{Inference for precision weighted average log risk ratio in meta-analysis of two-by-two tables}
\usage{
pwa_lrr(
  ai,
  n1i,
  ci,
  n2i,
  alternative = "two.sided",
  conf.level = 0.95,
  correction_factor = 0.01,
  ...
)
}
\arguments{
\item{ai:}{number of responders in group 1}

\item{n1i:}{sample size of group 1}

\item{ci:}{number of responders in group 2}

\item{n2i:}{sample size of group 2}
}
\description{
Inference for precision weighted average log risk ratio in meta-analysis of two-by-two tables
}
