% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-data.R
\name{pwa_meta}
\alias{pwa_meta}
\title{Inference for precision weighted average estimator in general meta-analysis}
\usage{
pwa_meta(yi, vi, ni, alternative = "two.sided", conf.level = 0.95)
}
\arguments{
\item{yi:}{study effect size}

\item{vi:}{variance estimate}

\item{ni:}{study sample size}
}
\description{
Inference for precision weighted average estimator in general meta-analysis
}
