% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{pwa_proportion}
\alias{pwa_proportion}
\title{Inference for precision weighted average proportion in proportion meta-analysis}
\usage{
pwa_proportion(
  xi,
  ni,
  alternative = "two.sided",
  conf.level = 0.95,
  correction_factor = 0.01,
  ...
)
}
\arguments{
\item{xi:}{number of responders in each study}

\item{ni:}{study sample size}
}
\description{
Inference for precision weighted average proportion in proportion meta-analysis
}
